requirements.txt
# requirements.txt
ollama
langchain
chromadb
tortoise-orm
aiosqlite
python-dotenv
langchain-community
langchain-ollama
langchain-chroma